Ceilometer RBAC Support for API
This proposal adds Role Based Access Control for the Ceilometer
API. Using RBAC is possible to restrict/enforce limitations in
the specific methods that a user can call. For instance it
would be possible to have only users with special roles in
creating or deleting alarms whilst all the users can verify
them.

A policy.json file will contain the specific rules to apply
for the different roles leaving to the deployer the ability
of enforcing specific restrictions without having to change
the implementation.

Co-Authored-By: Fabio Giannetti <fabio.giannetti@hp.com>

Change-Id: I965a5dc2143b89018b7bdfe48db440e35b749b6a