Unify the timestamp of samples generated by polling agents
Many of the pollsters define the timestamp individually for each generated
sample, which should really be timestamping based on when the data was polled
and not when each sample is generated. We need to set the timestamp of the
polled data to the timestamp when the polling starts for unity.

Related-Bug: #1491509
Blueprint unify-timestamp-of-polled-data

Change-Id: If22980c9fdf81e1c76dcd4a9053daf88d7f89d18